refactored text creation/editing into smaller units
