+++ JHotDraw/src/main/java/org/jhotdraw/draw/TextEditingTool.java
@@ -87,30 +87,8 @@ public class TextEditingTool extends AbstractTool implements ActionListener {
             if (newText.length() > 0) {
                 typingTarget.setText(newText);
             }
-            UndoableEdit edit = new AbstractUndoableEdit() {
-
-                @Override
-                public String getPresentationName() {
-                    ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.draw.Labels");
-                    return labels.getString("attribute.text.text");
-                }
-
-                @Override
-                public void undo() {
-                    super.undo();
-                    editedFigure.willChange();
-                    editedFigure.setText(oldText);
-                    editedFigure.changed();
-                }
-
-                @Override
-                public void redo() {
-                    super.redo();
-                    editedFigure.willChange();
-                    editedFigure.setText(newText);
-                    editedFigure.changed();
-                }
-            };
+            
+            UndoableEdit edit = new TextUndoableEdit(editedFigure, oldText, newText);
             getDrawing().fireUndoableEditHappened(edit);
 
             typingTarget.changed();
@@ -118,7 +96,6 @@ public class TextEditingTool extends AbstractTool implements ActionListener {
 
             textField.endOverlay();
         }
-    //	        view().checkDamage();
     }
 
     @Override
